Corrigindo campos de opções

+++ database/migrations/2022_01_04_204539_create_user_dados_table.php
@@ -17,8 +17,8 @@ class CreateUserDadosTable extends Migration
             $table->id();
             $table->string('telefone');
             $table->date('data_nasc');
-            $table->integer('estado_civil');
-            $table->integer('sexo');
+            $table->tinyInteger('estado_civil');
+            $table->tinyInteger('sexo');
             $table->unsignedBigInteger('user_id')->unique();
             $table->timestamps();
